# openapi.yaml
openapi: 3.0.0
info:
  title: ConnectVerse API
  description: API for ConnectVerse, a unified messaging platform for modern teams.
  version: 1.0.0
servers:
  - url: /api
    description: Local API server
paths:
...
  /customers:
    get:
      summary: List Customers
      description: Get a list of customers with filtering and pagination.
      tags: [Customers]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: source
          in: query
          schema:
            $ref: '#/components/schemas/Channel'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CustomerStatus'
        - name: tag
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: A list of customers.
          content:
            application/json:
              schema:
                type: object
                properties:
                  customers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer

    post:
      summary: Create Customer
      description: Add a new customer to the system.
      tags: [Customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCustomer'
      responses:
        '201':
          description: Customer created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /customers/{customerId}:
    get:
      summary: Get Customer by ID
      description: Retrieve details for a specific customer.
      tags: [Customers]
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found.
    put:
      summary: Update Customer
      description: Update an existing customer's details.
      tags: [Customers]
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCustomer' # Can be same as create
      responses:
        '200':
          description: Customer updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found.
  /customers/{customerId}/status:
    put:
      summary: Update Customer Status
      description: Update the funnel stage for a customer.
      tags: [Customers]
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/CustomerStatus'
      responses:
        '200':
          description: Status updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /customers/{customerId}/deals:
    post:
      summary: Create Deal
      description: Create a new deal for a customer.
      tags: [Customers, Deals]
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewDeal'
      responses:
        '201':
          description: Deal created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deal'
...
components:
  schemas:
    Channel:
      type: string
      enum: [whatsapp, messenger, instagram, tiktok]
    CustomerStatus:
      type: string
      enum: [new, contacted, qualified, unqualified, demo, won]
    DealStatus:
      type: string
      enum: ['Won', 'Lost', 'In Progress']
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        avatarUrl:
          type: string
          format: uri
        joined:
          type: string
          format: date
        tags:
          type: array
          items:
            type: string
        channel:
          $ref: '#/components/schemas/Channel'
        status:
          $ref: '#/components/schemas/CustomerStatus'
        dealName:
          type: string
        dealHistory:
          type: array
          items:
            $ref: '#/components/schemas/Deal'
    NewCustomer:
      type: object
      required: [name, email, channel, status]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        channel:
          $ref: '#/components/schemas/Channel'
        status:
          $ref: '#/components/schemas/CustomerStatus'
    Deal:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          $ref: '#/components/schemas/DealStatus'
        amount:
          type: number
        closeDate:
          type: string
          format: date
    NewDeal:
      type: object
      required: [name, amount]
      properties:
        name:
          type: string
        amount:
          type: number
...